Better organizing of UPE methods
